Lab5_docker_docker_compose_makefile/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pandas==2.2.0
numpy==1.26.4
python-dotenv==1.0.0